ReferencesSwitcherForDotNet.Library/SolutionReferencesSwitcher.cs
using System;
using System.Collections.Generic;
using System.IO;
using System.Linq;
using Microsoft.Build.Construction;
using Microsoft.Build.Evaluation;

namespace ReferencesSwitcherForDotNet.Library
{
    public class SolutionReferencesSwitcher
    {
        public void Switch(string solutionFullPath)
        {
            var solution = SolutionFile.Parse(solutionFullPath);
            foreach (var solutionProject in solution.ProjectsInOrder)
            {
                if (File.Exists(solutionProject.AbsolutePath))
                    SwitchReferencesForProjectReferences(solutionProject, solution);
            }
            ProjectCollection.GlobalProjectCollection.UnloadAllProjects();
        }

        private static void AddProjectReference(ProjectInSolution matchedSolutionProject, ProjectInSolution solutionProject, Project project)
        {
            var metadata = GetMetadataForProjectReference(matchedSolutionProject);
            var relativePath = GetRelativePathForProjectReference(solutionProject, matchedSolutionProject);
            project.AddItem("ProjectReference", relativePath, metadata);
        }

        private static List<KeyValuePair<string, string>> GetMetadataForProjectReference(ProjectInSolution matchedSolutionProject)
        {
            return new List<KeyValuePair<string, string>>
                   {
                       new KeyValuePair<string, string>("Project", matchedSolutionProject.ProjectGuid),
                       new KeyValuePair<string, string>("Name", matchedSolutionProject.ProjectName)
                   };
        }

        private static string GetRelativePathForProjectReference(ProjectInSolution solutionProject, ProjectInSolution matchedSolutionProject)
        {
            var uri = new Uri(solutionProject.AbsolutePath);
            var uriRelativePath = uri.MakeRelativeUri(new Uri(matchedSolutionProject.AbsolutePath));
            return uriRelativePath.ToString().Replace("/", Path.DirectorySeparatorChar.ToString());
        }

        private static void SwitchReferencesForProjectReferences(ProjectInSolution solutionProject, SolutionFile solution)
        {
            var project = new Project(solutionProject.AbsolutePath);
            foreach (var referenceItem in project.Items.Where(x => x.ItemType == "Reference").ToList())
            {
                var matchedSolutionProject = solution.ProjectsInOrder.FirstOrDefault(x => x.ProjectName == referenceItem.EvaluatedInclude);
                if (matchedSolutionProject != null)
                    SwitchReferenceWithProjectReference(matchedSolutionProject, solutionProject, project, referenceItem);
            }
        }

        private static void SwitchReferenceWithProjectReference(ProjectInSolution matchedSolutionProject, ProjectInSolution solutionProject, Project project, ProjectItem referenceItem)
        {
            AddProjectReference(matchedSolutionProject, solutionProject, project);
            project.RemoveItem(referenceItem);
            project.Save();
        }
    }
}